app/requirements.txt
fastapi>=0.109.1
uvicorn>=0.11.7
python-dotenv
python-multipart>=0.0.18
requests>=2.32.0
openai
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
icalendar
starlette>=0.40.0
itsdangerous
PyJWT
pytz